+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- isArray validator can now validate items with validators.
+- isArray validator can now validate items with validators and multiple types.
 
 ## [0.5.0] - 2016-05-09
 ### Added