+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - isArray validator can now validate items with validators and multiple types.
 - `@Controller` and all route decorators do accept `RequestHandler` as middlewares.
+- Example code for simple middleware.
+- Example code for passportJS middleware.
 
 ## [0.5.0] - 2016-05-09
 ### Added